server routes updated

+++ server/routes/forgetpassword.js
@@ -12,27 +12,39 @@ router.post('/forgot', async(req,res)=>{
     User.findOne({email: req.body.email})
     .then(async(user)=>{
         if(!user){
-            return res.status(401).json({  // Unauthorized access - lacks valid authentication
+            res.status(401).send({  // Unauthorized access - lacks valid authentication
                 message: 'No User Found'
             })
         }
-        const token = new Token({
-            userID: user._id,
-            token: crypto.randomBytes(32).toString("hex")
-        })  
-        token.save()
-        .then(()=>{
-            res.status(201).json({
-                message: 'Email has been sent to reset the password'
+        else{
+            Token.findOne({userID: user._id})
+            .then(async(userToken)=>{
+                if(!userToken){
+                    const userToken = new Token({
+                        userID: user._id,
+                        token: crypto.randomBytes(32).toString("hex")
+                    })  
+                    userToken.save()
+                }
+                const url = `${process.env.BASE_URL}password/reset/${user._id}/${userToken.token}`;
+                await sendEmail(user.firstName, user.email, 'Reset Your Password', url)
+                .then(()=>{
+                    res.status(201).send({
+                        message: 'Email has been sent to reset the password'
+                    })
+                })
+                .catch(err=>{
+                    res.status(500).send({error: "Internal Server Error"})
+                })
             })
-        })
-        const url = `${process.env.BASE_URL}password/reset/${user._id}/${token.token}`;
-        await sendEmail(user.firstName, user.email, 'Reset Your Password', url)
-
+            .catch(err=>{
+                res.status(500).send({error: "Internal Server Error"})
+            })
+        }
     })
     .catch(err=>{
         // console.log('Error: ', err);
-        res.status(400).json({
+        res.status(400).send({
             error: 'Reset Password Link error'
         })
     })
@@ -41,39 +53,39 @@ router.post('/forgot', async(req,res)=>{
 // Route for Reset Password
 router.put('/reset/:id/:token', async(req,res)=>{
     if(req.body.confPassword !== req.body.newPassword)
-        return res.status(400).json({
+        res.status(400).send({
             message: 'Password does not match'
         })
     Token.findOne({UserID: req.params.id})
     .then(async(user) => {
         if(!user){
-            return res.status(401).json({  // Unauthorized access - lacks valid authentication
+            res.status(401).send({  // Unauthorized access - lacks valid authentication
                 message: 'No User Found'
             })
         }
-        if(!user.token || user.token !== req.params.token ) return res.status(400).json({message: 'No Token Found'})
+        if(!user.token || user.token !== req.params.token ) return res.status(400).send({message: 'No Token Found'})
         bcrypt.hash(req.body.confPassword,10, async(err,hash)=>{
             if(err){
                 console.log(err);
-                return res.status(500).json({error: 'Internal Server Error'}); // Internal Server Error
+                res.status(500).send({error: 'Internal Server Error'}); // Internal Server Error
             }
             await User.updateOne({_id: req.params.id}, {$set:{password: hash}})
             await Token.deleteOne({userID: req.params.id})
             .then(result=>{
                 console.log('Password changed')
-                return res.status(200).json({
+                res.status(200).send({
                     message: 'Password changed successfully'
                 })
             })
             .catch(err=>{
-                return res.status(200).json({
+                res.status(200).send({
                     error: 'Password does not changed successfully'
                 })
             })
         })
     })
     .catch(err => {
-        return res.status(200).json({
+        res.status(200).send({
             error: 'Invalid Request'
         })
     })
